Refactored types into a more organized hierarchy.
